+++ src/payloads/plan_payloads.rs
@@ -37,7 +37,7 @@ pub struct Plan {
     pub name: String,
     pub price: BigDecimal,
     pub data_limit: Option<f64>, // in GB
-    pub duration_days: u64,
+    pub duration_days: u32,
     pub description: Option<String>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
@@ -49,7 +49,7 @@ pub struct NewPlan {
     pub name: String,
     pub price: BigDecimal,
     pub data_limit: Option<f64>,
-    pub duration_days: u64,
+    pub duration_days: u32,
     pub description: Option<String>,
 }
